+++ ClassManagmentDB/dbo/StoredProcedures/spStudent_Update.sql
@@ -2,8 +2,8 @@
 	@Id int,
 	@FirstName nvarchar(50),
 	@LastName nvarchar(50),
-	@NationalCode int,
-	@Mobile int,
+	@NationalCode bigint,
+	@Mobile bigint,
 	@RegisterDate DATETIME2
 AS
 begin
