+++ ClassManagmentDB/dbo/Tables/Student.sql
@@ -3,8 +3,8 @@
 	[Id] INT NOT NULL PRIMARY KEY IDENTITY, 
     [FirstName] NVARCHAR(50) NOT NULL, 
     [LastName] NVARCHAR(50) NOT NULL, 
-    [NationalCode] INT NOT NULL, 
-    [Mobile] INT NOT NULL, 
+    [NationalCode] BIGINT NOT NULL, 
+    [Mobile] BIGINT NOT NULL, 
     [RegisterDate] DATETIME2 NOT NULL DEFAULT GETDATE()
 	
 )
